+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'rob@coward-family.net'
 license          'Apache 2.0'
 description      'Installs/Configures opsview agent'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.2.0'
+version          '0.2.1'
 
 %w(redhat centos windows).each do |os|
   supports os